Make `is_notebook` a utility function
